+++ src/assets/locales/es/validations.js
@@ -3,5 +3,10 @@ export default {
   image: 'formato de imagen no válido',
   confirm: 'Las contraseñas no son iguales',
   password: 'Permite mayúsculas, minúsculas, números, y _ - símbolos',
-  driver: 'No se ha seleccionado ningún conductor'
+  driver: 'No se ha seleccionado ningún conductor',
+  required: 'El campo es requerido',
+  requiredNumbers: 'El Campo solo puede contener Numeros',
+  NotSpaces: 'El campo no puede contener espacios',
+  requiredMaxTen: 'El campo debe tener exactamente 10 caracteres',
+  requiredMinTree: 'El campo debe tener al menos 3 caracteres'
 }